Fully Functional
Made every section fully functional. The only bug per se is whenever a message is received, python can't unpack the function given that it hasn't run yet. It only affects the console anyway so it shouldn't be too big of an issue. Other than that, I added functionality to also allow it to be used in any text channel, not just a direct message.